refactor: Rename RoutineList and move components to Routines folder
